+++ CLI tools/cluster_rcon.js
@@ -9,7 +9,6 @@ if(process.argv.length < 3){
 	process.exit(1)
 }
 
-const masterIP = `${config.masterIP}:${config.masterPort}`;
 const token = config.masterAuthToken
 let commandArray = process.argv;
 commandArray.shift();
@@ -24,14 +23,14 @@ let command = commandArray.join(" ");
 let killBiters = `/silent-command local surface=game.surfaces[1] for key, entity in pairs(surface.find_entities_filtered({force="enemy"})) do entity.destroy() end`;
 
 (async ()=>{
-	needle.post(masterIP+"/api/runCommand", {
+	needle.post(config.masterURL+"/api/runCommand", {
 		broadcast:true,
 		command,
 	}, needleOptionsWithTokenAuthHeader, (err, resp) => {
 		console.log(resp.body)
 	});
 	// setInterval(()=>{
-		// needle.post(masterIP+"/api/runCommand", {
+		// needle.post(config.masterURL+"/api/runCommand", {
 		// broadcast:true,
 		// command: killBiters,
 	// }, needleOptionsWithTokenAuthHeader, (err, resp) => {